[Refactor]: Refactor and Cleanup
