+++ MessageBroker/EventPublisher.cs
@@ -1,8 +1,6 @@
 ﻿using System.Text;
-using System.Text.Json.Serialization;
 using Newtonsoft.Json;
 using RabbitMQ.Client;
-using RabbitMQ.Client.Events;
 
 namespace MessageBroker;
