+++ PactNet.Provider/Repository/IStudentRepo.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using PactNet.Provider.Models;
 
 namespace PactNet.Provider.Repository;
